Rename file, sprinkle in the magic one line for matrixRTCSession. and remove comment block
